Exercise 4.21 done

+++ osa4/blogilista/controllers/blogs.js
@@ -54,9 +54,27 @@ blogsRouter.post('/', async (request, response, next) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+    const userId = decodedToken.id
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      return response.status(404).json({ error: 'blog not found' })
+    }
+
+    if (blog.user.toString() !== userId.toString()) {
+      return response.status(401).json({ error: 'unauthorized: not the blog owner' })
+    }
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+    } catch (error) {
+        next(error)
+  }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
